More batteryWatcher.sh refinements

+++ cap-rachel-configure.sh
@@ -1815,33 +1815,31 @@ while :; do
         if [[ $(cat /tmp/battery_connected_status 2>/dev/null) == 0 ]]; then
             echo "[!] Battery not connected, monitor stopped at $(date)" >> /var/log/rachel/battery.log
             echo "[+] Battery connected:  $(cat /tmp/battery_connected_status)" >> /var/log/rachel/battery.log 2>/dev/null
-            exit 0
+            exit 1
         else
             # If charge status is low (should be at/above 400 at 99% battery), battery connected, and last charge level is below 95%, then there is a possible bad battery and do not run script
             if [[ $(cat /tmp/chargeStatus 2>/dev/null) -lt 400 ]] && [[ $(cat /tmp/chargeStatus 2>/dev/null) -gt -275 ]] && [[ $(cat /tmp/batteryLastChargeLevel 2>/dev/null) -lt 95 ]]; then
                 let "badBattChk++"
                 if [[ $badBattChk -ge 6 ]]; then
-                    echo "[!] Possible bad battery, monitor stopped at $(date)" >> /var/log/rachel/battery.log
+                    echo "[!] Possible bad battery or charger, monitor stopped at $(date)" >> /var/log/rachel/battery.log
                     echo "[+] Battery connected:  $(cat /tmp/battery_connected_status)" >> /var/log/rachel/battery.log 2>/dev/null
                     echo "[+] Battery last charge level:  $(cat /tmp/batteryLastChargeLevel)" >> /var/log/rachel/battery.log 2>/dev/null
                     echo "[+] Battery charge status:  $(cat /tmp/chargeStatus)" >> /var/log/rachel/battery.log 2>/dev/null
                     echo "[+] Bad battery check #:  $(echo $badBattChk)" >> /var/log/rachel/battery.log 2>/dev/null
-                    exit 0
+                    exit 1
                 fi
             else
                 badBattChk=0
                 # If charging level is less then -200 (power not connected) and last charge level is below 3%, stop KA Lite and safely shutdown CAP 
-                if [[ $(cat /tmp/chargeStatus 2>/dev/null) -lt -200 ]]; then
-                    if [[ $(cat /tmp/batteryLastChargeLevel 2>/dev/null) -lt 3 ]]; then
-                        echo "[!] Low battery shutdown at $(date)" >> /var/log/rachel/battery.log
-                        echo "[+] Battery connected:  $(cat /tmp/battery_connected_status)" >> /var/log/rachel/battery.log 2>/dev/null
-                        echo "[+] Battery last charge level:  $(cat /tmp/batteryLastChargeLevel)" >> /var/log/rachel/battery.log 2>/dev/null
-                        echo "[+] Battery charge status:  $(cat /tmp/chargeStatus)" >> /var/log/rachel/battery.log 2>/dev/null
-                        echo "[+] Bad battery check #:  $(echo $badBattChk)" >> /var/log/rachel/battery.log 2>/dev/null
-                        kalite stop
-                        shutdown -h now
-                        exit 0
-                    fi
+                if [[ $(cat /tmp/chargeStatus 2>/dev/null) -lt -200 ]] && [[ $(cat /tmp/batteryLastChargeLevel 2>/dev/null) -lt 3 ]]; then
+                    echo "[!] Low battery shutdown at $(date)" >> /var/log/rachel/battery.log
+                    echo "[+] Battery connected:  $(cat /tmp/battery_connected_status)" >> /var/log/rachel/battery.log 2>/dev/null
+                    echo "[+] Battery last charge level:  $(cat /tmp/batteryLastChargeLevel)" >> /var/log/rachel/battery.log 2>/dev/null
+                    echo "[+] Battery charge status:  $(cat /tmp/chargeStatus)" >> /var/log/rachel/battery.log 2>/dev/null
+                    echo "[+] Bad battery check #:  $(echo $badBattChk)" >> /var/log/rachel/battery.log 2>/dev/null
+                    kalite stop
+                    shutdown -h now
+                    exit 0
                 fi
             fi
         fi
@@ -1865,7 +1863,7 @@ After=network.target
 
 [Service]
 User=root
-Restart=always
+Restart=on-success
 ExecStart=/root/rachel-scripts/batteryWatcher.sh
 
 [Install]
@@ -1891,10 +1889,6 @@ author      "Sam <sam@hackersforcharity.org>"
 # Events
 start on net-device-up IFACE=eth0
 stop on shutdown
-
-# Automatically restart process if crashed
-respawn
-respawn limit 20 5
 
 # Run the script!
 script
